delete auth.py
